ISSUE #1282: Better tests
