+++ config/services.yaml
@@ -221,9 +221,6 @@ services:
         factory: [ null, 'fromConfig' ]
         arguments: [ '@=app_config("daemon.cron", [])' ]
 
-    App\Infrastructure\Daemon\Cron\Cron:
-      class: 'App\Infrastructure\Daemon\Cron\SystemCron'
-
     # Autowiring of vendor classes
     League\Flysystem\FilesystemOperator: '@default.storage'
