+++ src/Infrastructure/Console/ConsoleOutputAware.php
@@ -15,7 +15,7 @@ trait ConsoleOutputAware
         $this->output = $output;
     }
 
-    protected function getConsoleOutput(): OutputInterface
+    public function getConsoleOutput(): OutputInterface
     {
         if (is_null($this->output)) {
             throw new \RuntimeException('Call setConsoleOutput() before getConsoleOutput()');
